The main function has been moved from Client class as singleton implementation
